Moving categoryCHarts to frontend/src/components/dashboard
